fixup! feat(config): Allow ranges in envlist
