+++ src/tox/config/source/ini_section.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
-from tox.config.loader.ini.factor import extend_factors
-from tox.config.loader.replacer import expand_ranges
+from tox.config.loader.ini.factor import extend_factors, expand_ranges
 from tox.config.loader.section import Section
